refactor: move all implementation to internal/core/, mirrorstack.go is facade
Addresses \"root should only have mirrorstack.go\" directive. The SDK root
directory now contains exactly one Go file: mirrorstack.go (~131 lines).
All real implementation moved under internal/core/ where it can grow per
concern without polluting the package root.

File moves (git-tracked renames preserve history):

  mirrorstack.go → internal/core/module.go
  mirrorstack_test.go → internal/core/module_test.go
  cron.go, cron_test.go → internal/core/
  db.go, describe.go, describe_test.go → internal/core/
  event.go, event_test.go → internal/core/
  mcp.go, mcp_test.go → internal/core/
  resources.go → internal/core/
  task.go, task_test.go → internal/core/

Package renamed from \"package mirrorstack\" to \"package core\" in all
moved files (sed-driven).

Root mirrorstack.go is now a thin re-export:

  type Config = core.Config           // type alias
  type Module = core.Module            // type alias — methods flow through
  type TaskHandler = core.TaskHandler
  type TaskOption = core.TaskOption

  func New(cfg Config) (*Module, error) { return core.New(cfg) }
  func Init(cfg Config) error { return core.Init(cfg) }
  func Start() error { return core.Start() }
  // ...~25 wrapper functions for the rest of the convenience API

Generic forwarding (Resolve[T], MCPTool[In,Out], MCPResource[Out]) uses
explicit generic wrappers since Go doesn't support aliasing generic
functions.

Public API unchanged:
  - ms.DB(ctx), ms.MCPTool(...), ms.DependsOn(...), etc. — all work identically
  - *ms.Module methods (m.DB, m.Start, etc.) work via type alias
  - ms.Config{...} literals work via type alias
  - Test files in internal/core have full access to unexported Module fields

go test ./... + go vet + examples/template build — all green.
