+++ internal/core/cron.go
@@ -1,4 +1,4 @@
-package mirrorstack
+package core
 
 import (
 	"net/http"
