+++ internal/core/mcp_test.go
@@ -1,4 +1,4 @@
-package mirrorstack
+package core
 
 import (
 	"context"
